# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TESIS - Sistema de Clasificación de Riesgo",
        "description": "Documentación de la API del sistema de tesis",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/clasificaciones": {
            "get": {
                "tags": [
                    "Clasificación"
                ],
                "summary": "Listar clasificaciones",
                "description": "Obtiene el listado de clasificaciones con filtros opcionales por establecimiento y función",
                "operationId": "da023f8968a20c301ecf53dfbc870cb4",
                "parameters": [
                    {
                        "name": "establecimiento_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "funcion_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de clasificaciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clasificación"
                ],
                "summary": "Crear una nueva clasificación",
                "description": "Crea una clasificación y ejecuta el modelo de Machine Learning según la función asociada",
                "operationId": "05209098487bf31781b05ba699135071",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "establecimiento_id",
                                    "actividad_economica_id"
                                ],
                                "properties": {
                                    "establecimiento_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "actividad_economica_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Clasificación creada correctamente"
                    },
                    "422": {
                        "description": "Error de validación o reglas de negocio"
                    }
                }
            }
        },
        "/api/clasificaciones/{id}": {
            "get": {
                "tags": [
                    "Clasificación"
                ],
                "summary": "Obtener una clasificación",
                "description": "Mostrar una clasificación por ID",
                "operationId": "64ca9e085230477ca6cb46bd3e800110",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clasificación encontrada"
                    },
                    "404": {
                        "description": "Clasificación no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clasificación"
                ],
                "summary": "Eliminar una clasificación (baja lógica)",
                "description": "Eliminar (baja lógica) una clasificación",
                "operationId": "fd71d08b5699e466395b9d028530b078",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Clasificación eliminada"
                    }
                }
            }
        },
        "/api/clasificaciones/{id}/restore": {
            "post": {
                "tags": [
                    "Clasificación"
                ],
                "summary": "Restaurar una clasificación eliminada",
                "description": "Restaurar una clasificación eliminada",
                "operationId": "4162023eaddcefb96308645dc2d8448f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clasificación restaurada correctamente"
                    }
                }
            }
        },
        "/api/cuestionarios": {
            "get": {
                "tags": [
                    "Cuestionario"
                ],
                "summary": "Listar respuestas de cuestionarios",
                "description": "Obtiene todas las respuestas de los cuestionarios con sus relaciones",
                "operationId": "259931c9151f81c714d88588357eded5",
                "responses": {
                    "200": {
                        "description": "Listado de respuestas del cuestionario"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cuestionario"
                ],
                "summary": "Registrar respuestas de un cuestionario",
                "description": "Registra las respuestas del cuestionario para una clasificación específica",
                "operationId": "84f086f649164048bbcc4e472c153f67",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clasificacion_id",
                                    "respuestas"
                                ],
                                "properties": {
                                    "clasificacion_id": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "respuestas": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "pregunta_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "nivel_satisfaccion_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cuestionario registrado correctamente"
                    },
                    "400": {
                        "description": "La clasificación ya respondió el cuestionario"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/cuestionarios/{clasificacionId}": {
            "get": {
                "tags": [
                    "Cuestionario"
                ],
                "summary": "Obtener respuestas de un cuestionario por clasificación",
                "description": "Obtener las respuestas de una clasificación",
                "operationId": "bbfecf75aaba2a29f93479d5ec915cb5",
                "parameters": [
                    {
                        "name": "clasificacionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuestas encontradas"
                    },
                    "404": {
                        "description": "No hay respuestas registradas"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ingresa tu token de autenticación Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}